ORI-xxx Enhance lustre/ldiskfs build system
Enhance the lustre/ldiskfs build system so it is more robust, flexible,
and consistent with lustre/zfs build system.  This change is being made
in the interest of standardizing the infrastructure around optional
Lustre backends osds.

This change does not effect the current behavior of the --with-ldiskfs,
--enable-ext4, or --with-ldiskfsprogs configure options.  However, it
does remove the obsolete --with-ldiskfs-inkernel configure option which
was only used by LLNL.  It also adds the --with-ldiskfs-obj configure
option which improves flexibility.  And the --enable-ldiskfs-build
configure option to support building against the lustre-ldiskfs-devel
package.  The behavior of these options is consistent with their zfs
counterparts, see commit 8c7266ca528f730d96e579a323e4c233446c3fda for
further details.

  --enable-ext4           enable ldiskfs build using ext4
  --enable-ldiskfs-build  enable ldiskfs configure/make
  --with-ldiskfs=path     set path to ldiskfs source
  --with-ldiskfs-obj=path set path to ldiskfs objects
  --with-ldiskfsprogs     use alternate names for ldiskfs-enabled e2fsprogs

Sample ./configure results when building lustre and ldiskfs using
the kernel-devel and kernel-debuginfo-common packages.

  checking whether to enable ldiskfs... yes
  checking ldiskfs source directory... /home/behlendo/src/git/lustre/ldiskfs
  checking ldiskfs object directory... /home/behlendo/src/git/lustre/ldiskfs
  checking ldiskfs module symbols... Module.symvers
  checking ldiskfs source release... 3.3.0
  checking whether to use ext3 or ext4 source... ext4
  checking whether to build ldiskfs... yes
  checking for /home/behlendo/src/git/lustre/ldiskfs/configure... yes
  checking ext4 source directory...  /usr/src/debug/.../fs/ext4
  checking for /usr/src/debug/.../fs/ext4/dir.c...  yes
  checking for /usr/src/debug/.../fs/ext4/file.c...  yes
  checking for /usr/src/debug/.../fs/ext4/inode.c...  yes
  checking for /usr/src/debug/.../fs/ext4/super.c...  yes
  checking for /usr/src/debug/.../fs/ext4/extents.c...  yes
  checking if ext4_ext_walk_space() takes i_data_sem... yes
  checking if LDISKFS_SINGLEDATA_TRANS_BLOCKS takes sb as argument... yes
  checking if ldiskfs_discard_preallocations defined... yes
  checking if ldiskfs_ext_insert_extent needs 5 arguments... yes

In the context of this change additional cleanup has been done and
all of the ldiskfs specific code relocated to lustre-build-ldiskfs.m4.

Note that this change moves us closer to supporting patchless Lustre
servers with ldiskfs.  Once the remaining kernel patches for Lustre
are dropped you will be able to build Lustre using the distribution
provided kernel-devel and kernel-debuginfo-common packages.

Change-Id: I579919234771dcbbc59f265673ca72277f6d82d2
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>