ORI-360 osd-zfs: xattr in SA iff dataset property xattr=sa is set
Only store xattrs in SAs if the dataset property xattr=sa is set.
The osd will be notified dynamically if this property changes via a
registered callback function.  The primary motivation here is to
facilitate testing the performance benefit of the xattr in SA
feature.  Also update mkfs.lustre to set xattr=sa so the osd will
continue to use this feature by default.